# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 63307f5ac97c70218d65d9e0
  title: Zuct
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /lecturers:
    post:
      x-skaffolder-id: 63308709c97c70218d65da75
      x-skaffolder-name: create
      x-skaffolder-id-resource: 633085b7c97c70218d65da5d
      x-skaffolder-resource: lecturers
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - lecturers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/lecturers"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                l_first_name:
                  type: string
                l_gender:
                  type: string
                l_last_name:
                  type: string
                l_middle_name:
                  type: string
                l_phone:
                  type: number
    get:
      x-skaffolder-id: 6330870ac97c70218d65da7a
      x-skaffolder-name: list
      x-skaffolder-id-resource: 633085b7c97c70218d65da5d
      x-skaffolder-resource: lecturers
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - lecturers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/lecturers"
  "/lecturers/{id}":
    delete:
      x-skaffolder-id: 6330870ac97c70218d65da79
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 633085b7c97c70218d65da5d
      x-skaffolder-resource: lecturers
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - lecturers
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 63308709c97c70218d65da76
      x-skaffolder-name: get
      x-skaffolder-id-resource: 633085b7c97c70218d65da5d
      x-skaffolder-resource: lecturers
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - lecturers
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/lecturers"
    post:
      x-skaffolder-id: 63308709c97c70218d65da73
      x-skaffolder-name: update
      x-skaffolder-id-resource: 633085b7c97c70218d65da5d
      x-skaffolder-resource: lecturers
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - lecturers
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/lecturers"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/lecturers"
  /programs:
    post:
      x-skaffolder-id: 63308700c97c70218d65da6b
      x-skaffolder-name: create
      x-skaffolder-id-resource: 633086dfc97c70218d65da66
      x-skaffolder-resource: programs
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - programs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/programs"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pr_description:
                  type: string
                pr_name:
                  type: string
    get:
      x-skaffolder-id: 63308700c97c70218d65da6f
      x-skaffolder-name: list
      x-skaffolder-id-resource: 633086dfc97c70218d65da66
      x-skaffolder-resource: programs
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - programs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/programs"
  "/programs/{id}":
    delete:
      x-skaffolder-id: 63308700c97c70218d65da70
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 633086dfc97c70218d65da66
      x-skaffolder-resource: programs
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - programs
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 63308700c97c70218d65da69
      x-skaffolder-name: get
      x-skaffolder-id-resource: 633086dfc97c70218d65da66
      x-skaffolder-resource: programs
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - programs
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/programs"
    post:
      x-skaffolder-id: 63308700c97c70218d65da6c
      x-skaffolder-name: update
      x-skaffolder-id-resource: 633086dfc97c70218d65da66
      x-skaffolder-resource: programs
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - programs
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/programs"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/programs"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    lecturers:
      x-skaffolder-id: 633085b7c97c70218d65da5d
      x-skaffolder-id-db: 63307f5ac97c70218d65d9e1
      x-skaffolder-id-entity: 633085b7c97c70218d65da59
      x-skaffolder-url: /lecturers
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        l_first_name:
          type: string
          x-skaffolder-id-attr: 633085b7c97c70218d65da5a
          x-skaffolder-type: String
          x-skaffolder-required: true
        l_gender:
          type: string
          x-skaffolder-id-attr: 6330860fc97c70218d65da60
          x-skaffolder-type: String
          x-skaffolder-required: true
        l_last_name:
          type: string
          x-skaffolder-id-attr: 633085b7c97c70218d65da5c
          x-skaffolder-type: String
          x-skaffolder-required: true
        l_middle_name:
          type: string
          x-skaffolder-id-attr: 633085b7c97c70218d65da5b
          x-skaffolder-type: String
        l_phone:
          type: number
          x-skaffolder-id-attr: 63308625c97c70218d65da62
          x-skaffolder-type: Number
      x-skaffolder-relations:
        _contain:
          x-skaffolder-id: 63308b8dc97c70218d65da98
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 633085b7c97c70218d65da59
          x-skaffolder-ent2: 633086dfc97c70218d65da64
      required:
        - _id
        - l_first_name
        - l_gender
        - l_last_name
    programs:
      x-skaffolder-id: 633086dfc97c70218d65da66
      x-skaffolder-id-db: 63307f5ac97c70218d65d9e1
      x-skaffolder-id-entity: 633086dfc97c70218d65da64
      x-skaffolder-url: /programs
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        pr_description:
          type: string
          x-skaffolder-id-attr: 63308766c97c70218d65da7c
          x-skaffolder-type: String
          x-skaffolder-required: true
        pr_name:
          type: string
          x-skaffolder-id-attr: 633086dfc97c70218d65da65
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - pr_description
        - pr_name
...